add multithread into lib
